+++ src/imio/dashboard/browser/overrides.py
@@ -67,7 +67,16 @@ class IDDocumentGenerationView(DocumentGenerationView):
         if IFacetedNavigable.providedBy(self.context):
             brains = getDashboardQueryResult(self.context)
             generation_context['brains'] = brains
-            generation_context['uids'] = [brain.UID for brain in brains]
+            if pod_template.use_objects:
+                wrapped_objects = []
+                for brain in brains:
+                    generation_context['uids'].append(brain.UID)
+                    obj = brain.getObject()
+                    helper = obj.unrestrictedTraverse('@@document_generation_helper_view')
+                    wrapped_objects.append((helper.context, helper))
+                generation_context['objects'] = wrapped_objects
+            else:
+                generation_context['uids'] = [brain.UID for brain in brains]
 
         generation_context.update(super(IDDocumentGenerationView, self)._get_generation_context(helper_view,
                                                                                                 pod_template))